requirements.txt
kubernetes==24.2.0
PyYAML==5.4.1